coded the first 2 tasks

+++ Task_2.c
@@ -7,7 +7,7 @@ of type struct node and the data member in each node composed of a positive numb
 After the tree is constructed, the main() function calls a user-defined function delete_tree() that
 takes a pointer to the rootnode of the tree as input argument and deletes all the nodes in the tree. 
 However, in this case, the main() function utilizes the delete_tree() function to delete all the nodes
-in the right subtree of the rootnode and it does this by calling the function with the rightPtr member
+in the right subtree of the rootnode and it does this by calling the function with the adddress of the rightPtr member
 of the rootnode as input argument.
 
 For instance, if data was inserted into this binary search tree in the following order: 
@@ -32,6 +32,7 @@ the data in the printed output is delimited using one whitespace character.
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <string.h>
 
 struct node
 {
@@ -44,22 +45,25 @@ void insert_node(struct node** treePtr, int data);
 void inOrder(struct node* treePtr);
 void delete_tree(struct node** treePtr);
 
-int main() {
-	int temp = 0;
+int main(int argc, char* argv[]) 
+{
+	char* tokenPtr = NULL;
 	struct node* treePtr = NULL;
-    printf("Enter the value of the new data member: ");
-	scanf("%d", &temp);
-    while (temp > 0)
+    //printf("Enter the value of the new data member: ");
+	//scanf("%d", &temp);
+	tokenPtr = strtok(argv[1],",");
+    while (tokenPtr != NULL)
     {
-        insert_node(&treePtr, temp);
-        printf("Enter the value of the new data member: ");
-        scanf("%d", &temp);            
+        insert_node(&treePtr, atoi(tokenPtr));
+        // printf("Enter the value of the new data member: ");
+        // scanf("%d", &temp);  
+		tokenPtr = strtok(NULL,",");         
     }
-    printf("Initial version of binary tree:\n");
-    inOrder(treePtr);
-    printf("\n");
+    // printf("Initial version of binary tree:\n");
+    // inOrder(treePtr);
+    // printf("\n");
     delete_tree(&(treePtr->rightPtr));
-    printf("Modified version of binary tree:\n");
+    // printf("Modified version of binary tree:\n");
     inOrder(treePtr);
     printf("\n");
 }
@@ -97,9 +101,14 @@ void inOrder(struct node* treePtr)
 	}
 }
 
-void delete_tree(struct node** treePtr)
-{
-       free(*treePtr);
-	   delete_tree(&((*treePtr)->leftPtr));
-       delete_tree(&((*treePtr)->rightPtr));
+void delete_tree(struct node** treePtr) //uses postorder traversal (LRV)
+{	
+	if ((*treePtr) != NULL) //check if node exists
+	{
+		delete_tree(&((*treePtr)->leftPtr));
+		delete_tree(&((*treePtr)->rightPtr));
+		free(*treePtr);
+		(*treePtr) = NULL; //inform the tree that the node has been freed 
+	}
 }
+
